feat: fix backgroung file and login button state
